Add early invalidation to cacheExchange
Fix #58

Suppose a Client component fetches some data and is then unmounted, its
response will still remain cached. When a mutation is then performed
that invalidates its data, the component is not mounted to invalidate
this response itself. Instead it will render with the stale, cached
response.

To prevent this change introduces a secondary cache control. This
control keeps track of typenames as well in the cacheExchange. It maps
typenames to operation keys (which are also cache keys, woohoo!). When
a mutation then invalidates a couple of typenames in the cache, the
cacheExchange can go ahead and look up all cache keys that should be
invalidated, all while the Client component can stay unmounted.