Refactor subscription exchange for Wonka
We now assume the subscription-transport-ws format more strictly.
We expect an observable-ish structure to be returned and subscribe
to it. We then transform the ExecutionResult to the ExchangeResult
correctly to keep all assumptions intact.

This now also drops the operationName conversion, since it's optional
in subscription-transport-ws and most other transport layers.

It also now implements the unsubscription via the teardown signal.