Make composeExchanges more generic
It'd make sense for a user to expect a compose
function like composeExchanges to return the
same type as the input values. In this
case it'd make sense to receive Exchange not
ExchangeIO as a return value.

This would allow an external library to use
this utility to also bind some exchanges together
to expose as a single one.

The new signature of composeExchanges hence
accepts an array of exchanges and returns
a single one.

The fallbackExchangeIO has been separated and
is now passed in on the client.

The client now also only has the client.exchange
property for the composed exchange. This makes
sense because:

- Changing the exchanges wouldn't accomplish
  anything
- We don't need the property and it isn't
  particularly useful
- Exposing the composed exchange pipeline allows
  it to be potentially reused in a meaningful way