Add requestPolicy option to specify cache-usage
We've previously eliminated the `refetch` feature
as it didn't quite fit in with an otherwise pretty
minimal API.

As an alternative it's now possible to specify
the cache strategy, our default being "cache-first".

Additionally "cache-only", "network-only", and
"cache-and-network" are now supported.

This is quite a flexible approach that can also
be easily adopted by third-party cache exchanges
without seeming out-of-place.

And for our refetch queries the user can now
specify "network-only" or "cache-and-network"
to refetch a given query from network.