Add useRequest hook to simplify logic
Currently we manually rely on request.key in
useQuery to ensure that we don't trigger a
new effect, when it's not necessary, which
is error-prone.

Instead this hook uses useMemo and checks
request.key against the last request it generated.
If the key hasn't changed it returns the previous
reference.

This ensures reference equality, which removes
any check of request.key in useQuery