Add useImmediateEffect hook
We want to execute an effect immediately in render
on initial mount for suspense / SSR.

This is a very simple hook that does this by
abstracting effects in three phases:

- Initial mount (render)
- After mount
- Render

So on initial mount it will execute the effect
immediately. Then it will prevent the first
run of useEffect since the effect has already
been called, then the effect will run as usual.