Ensure that state changes are applied immediately on mount (#256)
* Ensure that setState are immediate on mount

It seems that in the current version React does
trigger *two* renders in our case.

We call executeQuery immediately which has a couple
of setStates. These are batched, it seems, but
do trigger a second update of the underlying
component, which is a shame.

Instead, we know that the initial executeQuery can
just safely merge its results into the initial
state, if it's during the initial mount.

This is why useImmediateState is added, which does
just that. Not the nicest way to solve this, but
the safest until React ships batched updates by
default.

* Improve comments on useImmediate* hooks

* Add tests for useImmediateState

* Simplify useImmediateEffect and add additional test

* Add useImmediateEffect to eslint react-hooks/exhaustive-deps

This adds the additionalHooks option to eslint-plugin-react-hooks
that will run the exhaustive-deps check on useImmediateEffect as
if it was just the regular useEffect hook.

* Update useMutation and useSubscription with useImmediateState

* Add missing useCallback to useMutation