Implement toSuspenseSource for React Suspense
This accepts a Wonka source and throws a promise
if the result is not resolving immediately. This
is done for suspensen and will also limit the
source to a single result.

If the source pushes a value synchronously it
is let through and won't throw a promise, which
also will subsequently let through more values
from the source.