Add queue to client's operation dispatching
We run the entire exchange pipeline synchronously,
since that's predictable, which is how Wonka's
sources work.

When an exchange dispatches an operation, for
instance with reexecuteOperation, this operation
is hence run immediately, which can be confusing
and lead to state that disagrees with itself,
since the operation is essentially then "nested".

Instead what we want is to queue operations up
if they're dispatched nestedly, so from inside
an exchange, and flush this queue after the
original dispatch has completed.