Refactor queries to allow for cascaded partial results (#59)
* Add isListNullable to SchemaPredicates

This is to check whether a given type for
a list may have nullable entries.

* Reimplement query operations using cascading nullables

First of all, this gets rid of `readOperation`, because
it's almost identical to `query`, which should save us
some bytes.

`readSelection` has been reimplemented. Instead of writing
fields onto data directly, it now first stores them in
`dataFieldValue`.

At any point when we encounter an uncached field that
is NOT nullable, we return undefined.
This is done by checking `dataFieldValue` against
the schema. This is now done in one place to save
some space and the hassle of reading repetitive code.

This works because we now represent uncached values
as `undefined`, which has been done in the past,
but has been removed. Hence:

- When `dataFieldValue` is `undefined` and a field
  is nullable then we write `null` and continue
- When `dataFieldValue` is `undefined` and a field
  is NOT nullable then we return `undefined` instead
  of `data` immediately
- Otherwise we write the field as usual

This is nice, because that means `undefined` cascades
upwards and this same logic also applies to resolved
links!

* Restore pruning root Query when no field was set

When no field has been set on Query and the result
would've been partially complete, then return
`undefined` for the entire Query.

* Update cachExchange to use partial instead of completeness

* Update Store to use read instead of startQuery

* Fix up tests using completeness to use partial

* Fix cacheOutcome to correctly be partial instead of full

* Fix readSelection value setting not respecting schema+cached

* Fix hasFields being accidentally set for nullable, uncached fields

* Fix assumed null for cache resolvers

* Correct invalidateQuery having an edge case where an empty query is returned

* Defer setting ctx.partial in readSelection

If we encounter an uncached, nullable field we'd set
ctx.partial, but if we then later return early with
undefined, that'd still be set!

Instead we need to set it after iterating, so that
early `return undefined` statements skip setting
ctx.partial

* Fix condition for using readRoot instead of readSelection

* Fix missing null field in mutation mock in Exchange tests

* Fix resolver links not being trusted when undefined

* Check for nullable entities in lists

* Extract shared getField code in SchemaPredicates

* Remove .only from query test