RFC: Remove impossible case of scalars where links are expected (#142)
* Remove case where scalar instead link is encountered

We remove this case because it's technically impossible
to encounter it. We can instead abort traversal at certain
points and not cache those values (readSelection modification).

We have a small identity function that will ensure that
TypeScript doesn't complain, which we can add warnings to
if we do want to check for this case.

* Remove tests related to this case

* Fix up broken tests

* Remove obsolete help.md entry

* Fix incorrect data in some benchmark cases

* Normalise `undefined` to `null` in ensureData

This prevents weird errors (of undefined accesses)
and kind of restores the prior behaviour without
the warning we used to have.