Fix cascading cancellations from components
wonka@^4.0.0 has some improved behaviour that fixed
some edge cases and bugs. As it turned out urql was
relying on one of them.
The switchMap in the hooks would cancel the last
query and start the next. However, this cancellation
cascades not only up to executeQuery to end and
teardown the operation, it also flows up through the
exchange pipeline, ending the entire thing like an
electrical surge.

This can be prevented by publishing the results
chain and making sure it can never be interrupted.