Modify subscribe in useSource not to allow getCurrentValue in effects
This stops us from creating temporary subscriptions when useSubscription
calls getCurrentValue in useEffect. This may happen because React first
checks whether it has missed any values.

Due to this change we may miss changed values, but this shouldn't be
an issue, because `subscribe` is pretty reliable in returning up-to-date
state.