Fix operation queue to only affect reexecuteOperation
Previously any call that dispatches an operation could
lead to a queued dispatch, which stops peeking from working
since the effect (exchange pipeline) isn't actually executed
on the spot. In such a case the queue is flushed later which
can lead to undefined behaviour, e.g. looping onOperationEnd,
which means that we think a query is active while it's not.

Instead we want to apply the operation queuing only to
client.reexecuteOperation, which is what's called from inside
exchanges.

The new behaviour: We now queue up operations from
reexecuteOperation when we're currently dispatching another
operation. Otherwise every operation executes immediately.