Update transform-debug-target transform
dispatchDebug cannot be undefined anymore (but will be a noop).

We can simply wrap the call expression in a ternary and let Terser/Closure
take care of eliminating the dead variables and assignments. This also
results in a lower chance of having a dead function stick around, which was
defined in the ObjectProperty transform.