Shift mutation results to top layer above pre-reservation
This makes a slight adjustment to the data layer reservation.
Optimistic Mutation layers are now not only cleared but will be
shifted to the front of the optimistic order queue when they're
written with a concrete result (non optimistic result).

The optimistic mutation layer is therefore still discarded but
the layer itself is reused, which allows for slight adjustments
in how layers are handled when they transition from being
optimistic to non-optimistic.

This also makes an adjustment that allows layers to transition
from being non-optimistic to optimistic. This shouldn't happen
in practice, but protects us from race-conditions.