Fix timing for out-of-band client.reexecuteOperation calls
When an operation becomes asynchronous, by matter of an exchange
delaying it, while it also calls client.reexecuteOpoeration, we
were calling client.reexecuteOperation out-of-band, which causes
the reexecuted operation to be dispatched immediately.
We can't detect the intent of the exchange, and don't know that it
may afterwards immediately send a result synchronously after the delay,
but we do know that once an operation becomes asynchronous, the timing
of client.reexecuteOperation is not sensitive anymore. So instead we
now choose to defer it by a tick.