Replace scheduleTask helper with inline Promise.resolve()
The scheduleTask helper was useful because globally we can't
tell whether `Promise` has been polyfilled for IE11 yet.
However, where this is called we _can_ already be sure of
it being polyfilled, which we do assume in other places
of the codebase.
Therefore it's safe to replace the `scheduleTask` polyfill
with a simple inline `Promise.resolve().then(fn)` call, which
is also more obvious.