(graphcache) - Clear optimistic updates before retrying in offlineExchange (#1080)
* Test for repeated optimistic update

* Enforce that all optimistic updates are atomic

NOTE: This changes an earlier (but rare) case's behaviour!
This is intentional, but wanted. We expected that this changes
how a "toggling mutation" may look in the UI. Specifically it
will never be able now to actually toggle as we go. That's
probably a good thing however as between a bad network
connection and going offline/online we can't tell whether
all toggling mutations will be delivered and applied in-order
and atomically.

* Add changeset

* Issue teardowns for failed mutations before retrying them

* Revert "Enforce that all optimistic updates are atomic"

This reverts commit b24dfc50f249f8b679b3ebb1ee31c130d257fdb5.

* Update test for new offlineExchange behaviour