(graphcache) - Increase consistency of __typename in cached results (#1185)
* Add `__typename` field as instructed by selection sets

The `__typename` field does not need to be added to root
types by default and must not be added first, as it shouldn't
be necessary. Instead we can assume that it's in the result
and copy it over once it's reached in the selection iteration.

* Update testing fixtures for query.test.ts

* Add changeset

* Update various test suites

* Restore skipped __typename field on write

* Update __typename field assignment to add alias

* Format documents before reading from cache in Store

* Add warning for writing selection sets without __typename

* Check test suites for silent warnings

* Update changeset

* Update Changeset to minor bump

I believe this should be a minor bump since we're adding a new
warning.