Add compatibility for Wonka compiled with ReScript >= 8
ReScript has new variant encoding from version 8 onwards.
This means that if Wonka is compiled with ReScript 8 or newer then our
manual encoding of a variant passed to `sink` doesn't work. To remedy
this we add the new variant encoding next to the old one. ReScript will
ignore the encoding it doesn't need so this should with versions of
ReScript before and after the change.