Add Vite-specific installation instructions
This is a common error that developers encounter when using either this or other GraphQL clients together with Vite. The questions and threads opened on this are many, but the answers are sparse and unhelpful, therefore a mention in the official documentation will help many people in the long term.