filter network-only requests from the ssrExchange
On the prior lines we defer clearing the data by a microtick which means that the redispatched "network-only" request will also be served from the ssr-cache. We could also opt to defer the redispatch by a microtick but imo it's better to treat network-only requests as bypassing this. Another possible solution would be to check whether we've seen this before as a `revalidation`