fix(graphcache): Switch fragment heuristic to always be truthy on writes
Previously we assumed that if an unknown fragment was being applied it'd
be an exceptional case and would not have to apply if the cache doesn't
already know all its fields.

We can now (theoretically) safely switch to always return `true` on writes.
When we encounter a field that isn't in the data, as returned by the API
then we can eagerly write anyway and let it fail.

This is important for cases where the fragment applies to an interface
of a node rather than to a concrete type.