Replace error fallback with text/* special content handling
The outer `catch` was still responsible for handling errors,
so we can avoid duplication here and still use `statusText`.
We should instead add a separate case for `text/*` contents.
Those should never really be returned by endpoints, but if they
are we can return them as it was in this branch before.