docs:refactored examples using the bang operator
## Summary

Using the bang operator (!) on non-boolean values can lead to unexpected results. For example, in the case of [this example](/basics/react-preact.md#pausing-usequery), using the `pause` option inside the `useQuery` hook, one might set the value of the `from` variable to be `0`. In that case the checks would be inaccurate since `!0 === true`. However, we only expect to pause the query only when `from` or `limit` is nullish. We can replace that condition with a more robust one.

Yes, it's an edge case, but new developers reading this documentation might not be aware of the presence of this bug and might have some unexpected results using this example in their code.

## Set of changes

Replaced `!from || !limit` with the more accurate `from === undefined || from === null || limit === undefined || limit === null` in the `useQuery` hook/composable on [React/Preact Basics](/basics/react-preact.md) and [Vue basics](/basics/vue.md).
An intermediate variable `shouldPause` got used to avoid inlining that long condition into the `pause` option.