Replace workaround with loc source check (See #1315)
Previously, we added the operationName of the current document's AST node
to the output string to guarantee a unique key. This was necessary
specifically for `graphql-tag/loader` which would preserve an incorrect
`loc.source` string with all operations, which then led to duplicate
keys in our own key calculations.

Instead, we can also distrust any `loc.source` strings and check for our
magic key to be in there.