Update retryExchange.ts
https://formidable.com/open-source/urql/docs/advanced/retry-operations/#:~:text=don%27t%20want%20to-,infinitely%20attempt,-an

The documentation implies that not passing in a `maxNumberAttempts` will result in retries occurring indefinitely. This is actually behaviour that I'm trying to get, but can't get because the code is not actually doing that - it defaults to `2`.

This simple modification allows for that behavior and is also consistent with the documentation.