Do not overwrite application provided abort signals
If the application has provided an abort signal to `fetchOptions` when
initializing the `Client`, then that abort signal shouldn't be
overwritten when `makeFetchSource` sets up its own abort controller.

In my case, this will allow me to easily set timeouts on requests from
my web server to my graphql server.

---

I wanted to test this, but I can't see how without intrusive changes to
the test suite.

Now we're using `AbortSignal.any` it validates that the arguments are
actually abort signals. The mock functions are no longer sufficient. But
the best way of getting an abort signal is through AbortControllers.
But the test suite removes access to AbortController in a couple
different files. For example here
https://github.com/urql-graphql/urql/blob/92a101a5497040641306a783b56d5fc3ee0917b1/scripts/vitest/setup.js#L2