Swap away from using AbortSignal.any
There is however a potential memory leak here. If one user supplied signal
is repeatedly reused, it will prevent the `abortController`s from being
garbage collected. To deal with this, we need to unsubscribe once the
fetch has completed. But how that can be done at this point is unclear.

Perhaps the `loadend` event would be appropriate.